+++ src/main/java/dev/quarris/ppfluids/client/screen/FluidFilterScreen.java
@@ -1,18 +1,17 @@
 package dev.quarris.ppfluids.client.screen;
 
 import de.ellpeck.prettypipes.pipe.containers.AbstractPipeGui;
-import dev.quarris.ppfluids.container.FluidRetrievalModuleContainer;
+import dev.quarris.ppfluids.container.FluidFilterContainer;
 import net.minecraft.client.gui.components.AbstractWidget;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.player.Inventory;
 
-public class FluidRetrievalModuleScreen extends AbstractPipeGui<FluidRetrievalModuleContainer> {
+public class FluidFilterScreen extends AbstractPipeGui<FluidFilterContainer> {
 
-    public FluidRetrievalModuleScreen(FluidRetrievalModuleContainer screenContainer, Inventory inv, Component titleIn) {
+    public FluidFilterScreen(FluidFilterContainer screenContainer, Inventory inv, Component titleIn) {
         super(screenContainer, inv, titleIn);
     }
 
-    @Override
     protected void init() {
         super.init();
         for (AbstractWidget button : this.menu.getFilter().createButtons(this, this.leftPos + this.imageWidth - 7, this.topPos + 17 + 32 + 20)) {
@@ -20,6 +19,5 @@ public class FluidRetrievalModuleScreen extends AbstractPipeGui<FluidRetrievalMo
         }
 
         this.addRenderableWidget(this.menu.getSelector().getButton(this.leftPos + 7, this.topPos + 17 + 32 + 20));
-
     }
 }