+++ src/main/java/dev/quarris/ppfluids/container/FluidLimiterContainer.java
@@ -1,0 +1,20 @@
+package dev.quarris.ppfluids.container;
+
+import de.ellpeck.prettypipes.pipe.containers.AbstractPipeContainer;
+import dev.quarris.ppfluids.item.FluidLimiterModuleItem;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.MenuType;
+import org.jetbrains.annotations.Nullable;
+
+public class FluidLimiterContainer extends AbstractPipeContainer<FluidLimiterModuleItem> {
+
+    public FluidLimiterContainer(@Nullable MenuType<?> type, int id, Player player, BlockPos pos, int moduleIndex) {
+        super(type, id, player, pos, moduleIndex);
+    }
+
+    @Override
+    protected void addSlots() {
+
+    }
+}