+++ src/main/java/dev/quarris/ppfluids/item/FluidModuleItem.java
@@ -6,6 +6,7 @@ import de.ellpeck.prettypipes.pipe.PipeBlockEntity;
 import net.minecraft.core.Direction;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.neoforged.neoforge.fluids.FluidStack;
 import net.neoforged.neoforge.fluids.capability.IFluidHandler;
 
 public class FluidModuleItem extends ModuleItem implements IFluidModule {
@@ -15,18 +16,18 @@ public class FluidModuleItem extends ModuleItem implements IFluidModule {
     }
 
     @Override
-    public boolean canNetworkSee(ItemStack module, PipeBlockEntity pipe, Direction dir, IFluidHandler destination) {
+    public int getMaxInsertionAmount(ItemStack module, PipeBlockEntity pipe, FluidStack fluidStack, IFluidHandler tank) {
+        return Integer.MAX_VALUE;
+    }
+
+    @Override
+    public boolean canNetworkSee(ItemStack module, PipeBlockEntity pipe, Direction dir, IFluidHandler tank) {
         return true;
     }
 
     @Override
-    public boolean canAcceptItem(ItemStack module, PipeBlockEntity pipe, ItemStack stack, Direction dir, IFluidHandler destination) {
+    public boolean canAcceptItem(ItemStack module, PipeBlockEntity pipe, ItemStack stack, Direction dir, IFluidHandler tank) {
         return true;
-    }
-
-    @Override
-    public int getMaxInsertionAmount(ItemStack module, PipeBlockEntity pipe, Direction dir, IFluidHandler destination) {
-        return 0;
     }
 
     @Override