+++ src/main/java/dev/quarris/ppfluids/registry/MenuSetup.java
@@ -5,9 +5,10 @@ import de.ellpeck.prettypipes.items.IModule;
 import de.ellpeck.prettypipes.pipe.PipeBlockEntity;
 import de.ellpeck.prettypipes.pipe.containers.AbstractPipeContainer;
 import dev.quarris.ppfluids.ModRef;
-import dev.quarris.ppfluids.container.FluidExtractionModuleContainer;
-import dev.quarris.ppfluids.container.FluidFilterModuleContainer;
-import dev.quarris.ppfluids.container.FluidRetrievalModuleContainer;
+import dev.quarris.ppfluids.container.FluidExtractionContainer;
+import dev.quarris.ppfluids.container.FluidFilterContainer;
+import dev.quarris.ppfluids.container.FluidLimiterContainer;
+import dev.quarris.ppfluids.container.FluidRetrievalContainer;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.ItemStack;
@@ -20,9 +21,10 @@ public class MenuSetup {
 
     public static final DeferredRegister<MenuType<?>> REGISTRY = DeferredRegister.create(Registries.MENU, ModRef.ID);
 
-    public static final DeferredHolder<MenuType<?>, MenuType<FluidFilterModuleContainer>> FLUID_FILTER_CONTAINER = REGISTRY.register("fluid_filter", MenuSetup::createPipeContainer);
-    public static final DeferredHolder<MenuType<?>, MenuType<FluidExtractionModuleContainer>> FLUID_EXTRACTION_CONTAINER = REGISTRY.register("fluid_extraction", MenuSetup::createPipeContainer);
-    public static final DeferredHolder<MenuType<?>, MenuType<FluidRetrievalModuleContainer>> FLUID_RETRIEVAL_CONTAINER = REGISTRY.register("fluid_retrieval", MenuSetup::createPipeContainer);
+    public static final DeferredHolder<MenuType<?>, MenuType<FluidFilterContainer>> FLUID_FILTER = REGISTRY.register("fluid_filter", MenuSetup::createPipeContainer);
+    public static final DeferredHolder<MenuType<?>, MenuType<FluidExtractionContainer>> FLUID_EXTRACTION = REGISTRY.register("fluid_extraction", MenuSetup::createPipeContainer);
+    public static final DeferredHolder<MenuType<?>, MenuType<FluidRetrievalContainer>> FLUID_RETRIEVAL = REGISTRY.register("fluid_retrieval", MenuSetup::createPipeContainer);
+    public static final DeferredHolder<MenuType<?>, MenuType<FluidLimiterContainer>> FLUID_LIMITER = REGISTRY.register("fluid_limiter", MenuSetup::createPipeContainer);
 
 
     public static void init(IEventBus bus) {